+++ packages/rpc_dart_analytics/lib/rpc_dart_analytics.dart
@@ -3,4 +3,4 @@ library rpc_dart_analytics;
 export 'src/analytics_client.dart';
 export 'src/config.dart';
 export 'src/messages.dart'
-    show AnalyticsStatusSnapshot;
+    show AnalyticsStatusSnapshot, AnalyticsDiagnosticsSnapshot, AnalyticsDiagnosticsEvent;